make customer and crafts man in a separate controller and service and rename files
