+++ src/main/kotlin/com/crafto/crafto_backend/controller/SharedController.kt
@@ -1,16 +1,15 @@
 package com.crafto.crafto_backend.controller
 
-import com.crafto.crafto_backend.response.CategoryResponse
-import com.crafto.crafto_backend.service.CraftoService
+import com.crafto.crafto_backend.dto.CategoryResponse
+import com.crafto.crafto_backend.service.SharedService
 import org.springframework.web.bind.annotation.*
 
 // GET  http://localhost:8085/categories
 
 @RestController
-@RequestMapping("/categories")
-class CraftoController(private val service: CraftoService) {
+class SharedController(private val service: SharedService) {
 
-    @GetMapping
+    @GetMapping("/categories")
     fun GetAllCategories(): List<CategoryResponse> {
         return service.getAllCategories()
     }
