+++ src/main/kotlin/com/crafto/crafto_backend/dto/CategoryResponse.kt
@@ -1,4 +1,4 @@
-package com.crafto.crafto_backend.response
+package com.crafto.crafto_backend.dto
 
 data class CategoryResponse(
     val categoryName:String,
