+++ src/main/kotlin/com/crafto/crafto_backend/repository/CategoryRepository.kt
@@ -3,4 +3,4 @@ package com.crafto.crafto_backend.repository
 import com.crafto.crafto_backend.entity.Category
 import org.springframework.data.mongodb.repository.MongoRepository
 
-interface CraftoRepository:MongoRepository<Category, String>
+interface CategoryRepository:MongoRepository<Category, String>
