+++ src/main/kotlin/com/crafto/crafto_backend/service/CraftsManService.kt
@@ -1,0 +1,8 @@
+package com.crafto.crafto_backend.service
+
+import org.springframework.stereotype.Service
+
+@Service
+class CraftsManService(){
+
+}
